refactor: Update expert deployment components IFix bug.  reorganize deployment components with new file structure and styling
