+++ src/app/modules/dashboard/pages/deployment/create-deployment/components/ai-assistant/ai-assistant.css
