+++ src/app/modules/dashboard/pages/deployment/create-deployment/components/mode-selector/mode-selector.html
