+++ src/app/modules/dashboard/pages/deployment/create-deployment/components/quick-deployment/quick-deployment.css
